Fixed MissingNo
Fixed MissingNo. The Font Sprite now loads correctly and the NewBox doesn't crash when interacting with it.
It's based on the Gengar Animation.
And it does render the images oddly in the Pokedex, but I think this is a minimal issue. It is the Glitch Pokemon, so I don't mind it being imperfect.